pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nsw-load-forecasting"
version = "0.1.0"
description = "End-to-end NSW load forecasting research framework"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.1",
  "pyyaml>=6.0",
  "scikit-learn>=1.4",
  "matplotlib>=3.8",
  "tqdm>=4.66",
  "torch>=2.2",
]

[tool.setuptools.packages.find]
where = ["src"]
